Attempting to handle paragraphs in one go.
This approach seems flawed because we aren't applying push-down automata. However, the near-term problem is foot-note processing inside the paragraph, which is wrong (we get unnested link and sup tags).

We may want to trace the footnote to its content, although apparently we can render them in-situ.

The next problem is that after these paragraphs we arrive at an <ul> for the footnotes. We will want to render these as Markdown lists eventually for body text but it isn't necessary for footnotes. So perhaps we should start examining the classes of the containing structure to make decisions about how to render.